requirements.txt
pandas
xlrd
openpyxl
streamlit
firebase-admin
PyJWT
pyrebase4
plotly

